Fixing output

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -7,7 +7,7 @@
 
 void print_number(int n)
 {
-	int temp, i, num, pow, exp;
+	long temp, i, num, pow, exp;
 
 	num = n;
 	exp = pow = 1;
